requirements.txt
# Web Framework
fastapi==0.115.5
uvicorn[standard]==0.32.1

# Database
sqlalchemy==2.0.36
psycopg2-binary==2.9.10

# Validation
pydantic==2.10.3
pydantic-settings==2.6.1
email-validator==2.2.0

# Security
passlib[bcrypt]==1.7.4
pyjwt==2.10.1

# Email
aiosmtplib==3.0.2

# Cache
redis==5.2.0
